Renamet course_tick, fixet id bug når courseAddLessons bliver lavet, og courseAddLesson sker efter et course lige er blevet lavet og bruger dens id

+++ lesson-app/src/main/java/dk/aau/cs/ds302e18/app/domain/CalendarViewModel.java
@@ -6,19 +6,19 @@ import java.util.Date;
 
 public class CalendarViewModel
 {
-    public int getId()
+    public long getId()
     {
         return id;
     }
 
-    public void setId(int id)
+    public void setId(long id)
     {
         this.id = id;
     }
 
-    private int id;
+    private long id;
 
-    public CalendarViewModel(int id, String color, String title, Date start)
+    public CalendarViewModel(long id, String color, String title, Date start)
     {
         this.id = id;
         this.color = color;
